Extract defineObservableProperty into its own file
