+++ app/javascript/packages/document-capture/components/acuant-capture-canvas.jsx
@@ -3,28 +3,7 @@ import { useContext, useEffect, useRef } from 'react';
 import { getAssetPath } from '@18f/identity-assets';
 import { useI18n } from '@18f/identity-react-i18n';
 import AcuantContext from '../context/acuant';
-
-/**
- * Defines a property on the given object, calling the change callback when that property is set to
- * a new value.
- *
- * @param {any} object Object on which to define property.
- * @param {string} property Property name to observe.
- * @param {(nextValue: any) => void} onChangeCallback Callback to trigger on change.
- */
-export function defineObservableProperty(object, property, onChangeCallback) {
-  let currentValue;
-
-  Object.defineProperty(object, property, {
-    get() {
-      return currentValue;
-    },
-    set(nextValue) {
-      currentValue = nextValue;
-      onChangeCallback(nextValue);
-    },
-  });
-}
+import { defineObservableProperty } from '../higher-order/observable-property';
 
 function AcuantCaptureCanvas() {
   const { isReady, acuantCaptureMode, setAcuantCaptureMode } = useContext(AcuantContext);
