+++ app/javascript/packages/document-capture/higher-order/observable-property.jsx
@@ -1,0 +1,21 @@
+/**
+ * Defines a property on the given object, calling the change callback when that property is set to
+ * a new value.
+ *
+ * @param {any} object Object on which to define property.
+ * @param {string} property Property name to observe.
+ * @param {(nextValue: any) => void} onChangeCallback Callback to trigger on change.
+ */
+export function defineObservableProperty(object, property, onChangeCallback) {
+  let currentValue;
+
+  Object.defineProperty(object, property, {
+    get() {
+      return currentValue;
+    },
+    set(nextValue) {
+      currentValue = nextValue;
+      onChangeCallback(nextValue);
+    },
+  });
+}
